CamundaUtils renamed to CamundaVerifier, test moved to new class CamundaVerifierTest
